reintegrado codigo de Upgrade to a10

See merge request kennethsomarriba96/frontend-v2!1

+++ src/app/services/main.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable, of, Subject } from 'rxjs';
+import { Observable, of, Subject, observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { wsModel } from 'src/app/models/ws.model';
 import { Api } from 'src/app/models/api.model';
@@ -26,7 +26,7 @@ export class MainService {
     return this.client.get<any>(this.getUrl());
 
   }
-  getByID(id: number|string): Observable<any> {
+  getByID(id: number | string): Observable<any> {
     return this.client.get<any>(`${this.getUrl()}${id}`);
 
   }
@@ -38,28 +38,31 @@ export class MainService {
     return this.client.post(this.getUrl(), body, head);
   }
 
-  update(body: any, id: string|number): Observable<any> {
+  update(body: any, id: string | number): Observable<any> {
     const head: any = {};
     head['Content-Type'] = 'application/json';
     return this.client.put(`${this.getUrl()}${id}`, body, head);
   }
 
-  delete(id: any): Observable<any> {
+  delete(id: any): Observable<any>{
     const head: any = {};
-    head['Content-Type'] = 'application/json';
-    return this.client.delete(this.getUrl() + id, head);
+    if (confirm('¿Esta seguro que desea eliminar?')) {
+      head['Content-Type'] = 'application/json';
+      return this.client.delete(this.getUrl() + id, head);
+    }
+    return new Observable();
   }
 
   getList() {
     return this.list$.asObservable();
   }
 
-  getByFiltro(filtro: string, id: string|number): Observable<any> {
-    console.log(`${this.getUrl()}${filtro}=${id}`);
+  getByFiltro(filtro: string, id: string | number): Observable<any> {
+    // console.log(`${this.getUrl()}${filtro}=${id}`);
     return this.client.get<any>(`${this.getUrl()}${filtro}=${id}`);
   }
 
-  getByPlan(query: string, filtro: string | number, id: string|number): Observable<any> {
+  getByPlan(query: string, filtro: string | number, id: string | number): Observable<any> {
     console.log(`${this.getUrl()}${query}=${filtro}/${id}`);
     return this.client.get<any>(`${this.getUrl()}${query}=${filtro}/${id}`);
   }
